+++ lib-common/src/main/java/io/boomerang/common/repository/TaskRunRepository.java
@@ -1,4 +1,4 @@
-package io.boomerang.engine.repository;
+package io.boomerang.common.repository;
 
 import io.boomerang.common.entity.TaskRunEntity;
 import java.util.List;
