+++ lib-common/src/main/java/io/boomerang/common/repository/WorkflowRepository.java
@@ -1,4 +1,4 @@
-package io.boomerang.engine.repository;
+package io.boomerang.common.repository;
 
 import io.boomerang.common.entity.WorkflowEntity;
 import io.boomerang.common.enums.WorkflowStatus;
