+++ service-engine/src/main/java/io/boomerang/aspect/WorkflowRunEntityUpdateInterceptor.java
@@ -4,8 +4,8 @@ import io.boomerang.audit.AuditInterceptor;
 import io.boomerang.audit.AuditType;
 import io.boomerang.common.entity.WorkflowRunEntity;
 import io.boomerang.common.enums.RunStatus;
+import io.boomerang.common.repository.WorkflowRunRepository;
 import io.boomerang.engine.EventSinkService;
-import io.boomerang.engine.repository.WorkflowRunRepository;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
