+++ service-engine/src/main/java/io/boomerang/engine/ParameterManager.java
@@ -12,8 +12,8 @@ import io.boomerang.common.enums.ParamType;
 import io.boomerang.common.model.ParamLayers;
 import io.boomerang.common.model.RunParam;
 import io.boomerang.common.model.RunResult;
-import io.boomerang.engine.repository.TaskRunRepository;
-import io.boomerang.engine.repository.WorkflowRunRepository;
+import io.boomerang.common.repository.TaskRunRepository;
+import io.boomerang.common.repository.WorkflowRunRepository;
 import io.boomerang.util.ParameterUtil;
 import java.util.ArrayList;
 import java.util.HashMap;
