+++ service-engine/src/main/java/io/boomerang/engine/TaskExecutionService.java
@@ -8,9 +8,9 @@ import io.boomerang.common.entity.WorkflowRunEntity;
 import io.boomerang.common.enums.*;
 import io.boomerang.common.model.*;
 import io.boomerang.common.model.WorkflowSchedule;
-import io.boomerang.engine.repository.ActionRepository;
-import io.boomerang.engine.repository.TaskRunRepository;
-import io.boomerang.engine.repository.WorkflowRunRepository;
+import io.boomerang.common.repository.ActionRepository;
+import io.boomerang.common.repository.TaskRunRepository;
+import io.boomerang.common.repository.WorkflowRunRepository;
 import io.boomerang.error.BoomerangError;
 import io.boomerang.error.BoomerangException;
 import io.boomerang.util.ParameterUtil;
