+++ service-flow/src/main/java/io/boomerang/Application.java
@@ -9,6 +9,7 @@ import io.swagger.v3.oas.annotations.security.SecuritySchemes;
 import java.time.Clock;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.web.bind.annotation.RestController;
@@ -34,6 +35,7 @@ import org.springframework.web.bind.annotation.RestController;
 @SpringBootApplication
 @EnableWebSecurity
 @RestController
+@EnableCaching
 public class Application {
 
   public static void main(String[] args) {
