+++ service-flow/src/main/java/io/boomerang/client/EngineClient.java
@@ -1,17 +1,12 @@
 package io.boomerang.client;
 
-import io.boomerang.common.model.ChangeLogVersion;
-import io.boomerang.common.model.Task;
 import io.boomerang.common.model.TaskRun;
 import io.boomerang.common.model.TaskRunEndRequest;
-import io.boomerang.common.model.Workflow;
-import io.boomerang.common.model.WorkflowCount;
 import io.boomerang.common.model.WorkflowRun;
 import io.boomerang.common.model.WorkflowRunCount;
 import io.boomerang.common.model.WorkflowRunInsight;
 import io.boomerang.common.model.WorkflowRunRequest;
 import io.boomerang.common.model.WorkflowSubmitRequest;
-import io.boomerang.common.model.WorkflowTemplate;
 import io.boomerang.error.BoomerangException;
 import io.boomerang.workflow.model.WorkflowRunEventRequest;
 import java.io.InputStream;
@@ -499,186 +494,6 @@ public class EngineClient {
    * ************************************** Workflow endpoints
    * **************************************
    */
-
-  public Workflow getWorkflow(String workflowId, Optional<Integer> version, boolean withTasks) {
-    try {
-      String url = getWorkflowURL.replace("{workflowId}", workflowId);
-      Map<String, String> requestParams = new HashMap<>();
-      if (version.isPresent()) {
-        requestParams.put("version", version.get().toString());
-      }
-      requestParams.put("withTasks", Boolean.toString(withTasks));
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      ResponseEntity<Workflow> response = restTemplate.getForEntity(encodedURL, Workflow.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public WorkflowResponsePage queryWorkflows(
-      Optional<Integer> queryLimit,
-      Optional<Integer> queryPage,
-      Optional<Direction> querySort,
-      Optional<List<String>> queryLabels,
-      Optional<List<String>> queryStatus,
-      Optional<List<String>> queryWorkflows) {
-    try {
-      UriComponentsBuilder urlBuilder = UriComponentsBuilder.fromHttpUrl(queryWorkflowURL);
-      if (queryPage.isPresent()) {
-        urlBuilder.queryParam("page", Integer.toString(queryPage.get()));
-      }
-      if (queryLimit.isPresent()) {
-        urlBuilder.queryParam("limit", Integer.toString(queryLimit.get()));
-      }
-      if (querySort.isPresent()) {
-        urlBuilder.queryParam("sort", querySort.get());
-      }
-      if (queryLabels.isPresent()) {
-        urlBuilder.queryParam("labels", queryLabels.get());
-      }
-      if (queryStatus.isPresent()) {
-        urlBuilder.queryParam("status", queryStatus.get());
-      }
-      if (queryWorkflows.isPresent() && !queryWorkflows.get().isEmpty()) {
-        urlBuilder.queryParam("workflows", queryWorkflows.get());
-      }
-      URI encodedURI = urlBuilder.build().encode().toUri();
-
-      LOGGER.info("Query URL: " + encodedURI);
-
-      ResponseEntity<WorkflowResponsePage> response =
-          restTemplate.getForEntity(encodedURI, WorkflowResponsePage.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().getContent().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public WorkflowCount countWorkflows(
-      Optional<List<String>> queryLabels,
-      Optional<List<String>> queryWorkflows,
-      Optional<Long> fromDate,
-      Optional<Long> toDate) {
-    try {
-      UriComponentsBuilder urlBuilder = UriComponentsBuilder.fromHttpUrl(countWorkflowURL);
-      if (fromDate.isPresent()) {
-        urlBuilder.queryParam("fromDate", fromDate.get());
-      }
-      if (toDate.isPresent()) {
-        urlBuilder.queryParam("toDate", toDate.get());
-      }
-      if (queryLabels.isPresent()) {
-        urlBuilder.queryParam("labels", queryLabels.get());
-      }
-      if (queryWorkflows.isPresent() && !queryWorkflows.get().isEmpty()) {
-        urlBuilder.queryParam("workflows", queryWorkflows.get());
-      }
-      URI encodedURI = urlBuilder.build().encode().toUri();
-
-      LOGGER.info("Query URL: " + encodedURI);
-
-      ResponseEntity<WorkflowCount> response =
-          restTemplate.getForEntity(encodedURI, WorkflowCount.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public Workflow createWorkflow(Workflow workflow) {
-    try {
-      LOGGER.info("URL: " + createWorkflowURL);
-
-      ResponseEntity<Workflow> response =
-          restTemplate.postForEntity(createWorkflowURL, workflow, Workflow.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public Workflow applyWorkflow(Workflow workflow, boolean replace) {
-    try {
-      String url = applyWorkflowURL;
-      Map<String, String> requestParams = new HashMap<>();
-      requestParams.put("replace", Boolean.toString(replace));
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      final HttpHeaders headers = new HttpHeaders();
-      headers.setContentType(MediaType.APPLICATION_JSON);
-      HttpEntity<Workflow> entity = new HttpEntity<Workflow>(workflow, headers);
-      ResponseEntity<Workflow> response =
-          restTemplate.exchange(encodedURL, HttpMethod.PUT, entity, Workflow.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
   public WorkflowRun submitWorkflow(
       String workflowId, WorkflowSubmitRequest request, boolean start) {
     try {
@@ -699,102 +514,6 @@ public class EngineClient {
       LOGGER.info("Content Response: " + response.getBody().toString());
 
       return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public List<ChangeLogVersion> getWorkflowChangeLog(String workflowId) {
-    try {
-      String url = changelogWorkflowURL.replace("{workflowId}", workflowId);
-
-      LOGGER.info("URL: " + url);
-
-      ResponseEntity<ChangeLogVersion[]> response =
-          restTemplate.getForEntity(url, ChangeLogVersion[].class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return Arrays.asList(response.getBody());
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public void enableWorkflow(String workflowId) {
-    try {
-      String url = enableWorkflowURL.replace("{workflowId}", workflowId);
-
-      LOGGER.info("URL: " + url);
-      ResponseEntity<Void> response = restTemplate.exchange(url, HttpMethod.PUT, null, Void.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-
-      if (!HttpStatus.NO_CONTENT.equals(response.getStatusCode())) {
-        throw new RestClientException("Unable to enable Workflow");
-      }
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public void disableWorkflow(String workflowId) {
-    try {
-      String url = disableWorkflowURL.replace("{workflowId}", workflowId);
-
-      LOGGER.info("URL: " + url);
-      ResponseEntity<Void> response = restTemplate.exchange(url, HttpMethod.PUT, null, Void.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-
-      if (!HttpStatus.NO_CONTENT.equals(response.getStatusCode())) {
-        throw new RestClientException("Unable to disable Workflow");
-      }
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public void deleteWorkflow(String workflowId) {
-    try {
-      String url = deleteWorkflowURL.replace("{workflowId}", workflowId);
-      Map<String, String> requestParams = new HashMap<>();
-      requestParams.put("cascade", Boolean.toString(false));
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-      ResponseEntity<Void> response =
-          restTemplate.exchange(encodedURL, HttpMethod.DELETE, null, Void.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
     } catch (RestClientException ex) {
       LOGGER.error(ex.toString());
       throw new BoomerangException(
@@ -896,353 +615,5 @@ public class EngineClient {
       }
       return null;
     };
-  }
-
-  /*
-   * ************************************** Task endpoints
-   * **************************************
-   */
-
-  public Task getTask(String ref, Optional<Integer> version) {
-    try {
-      String url = getTaskURL.replace("{ref}", ref);
-      Map<String, String> requestParams = new HashMap<>();
-      if (version.isPresent()) {
-        requestParams.put("version", version.get().toString());
-      }
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      ResponseEntity<Task> response = restTemplate.getForEntity(encodedURL, Task.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public TaskResponsePage queryTask(
-      Optional<Integer> queryLimit,
-      Optional<Integer> queryPage,
-      Optional<Direction> querySort,
-      Optional<List<String>> queryLabels,
-      Optional<List<String>> queryStatus,
-      List<String> queryRefs) {
-    try {
-      UriComponentsBuilder urlBuilder = UriComponentsBuilder.fromHttpUrl(queryTaskURL);
-      if (queryPage.isPresent()) {
-        urlBuilder.queryParam("page", Integer.toString(queryPage.get()));
-      }
-      if (queryLimit.isPresent()) {
-        urlBuilder.queryParam("limit", Integer.toString(queryLimit.get()));
-      }
-      if (querySort.isPresent()) {
-        urlBuilder.queryParam("sort", querySort.get());
-      }
-      if (queryLabels.isPresent()) {
-        urlBuilder.queryParam("labels", queryLabels.get());
-      }
-      if (queryStatus.isPresent()) {
-        urlBuilder.queryParam("status", queryStatus.get());
-      }
-      urlBuilder.queryParam("ids", queryRefs);
-      URI encodedURI = urlBuilder.build().encode().toUri();
-
-      LOGGER.info("Query URL: " + encodedURI);
-
-      ResponseEntity<TaskResponsePage> response =
-          restTemplate.getForEntity(encodedURI, TaskResponsePage.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().getContent().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public Task createTask(Task request) {
-    try {
-      LOGGER.info("URL: " + createTaskURL);
-
-      ResponseEntity<Task> response =
-          restTemplate.postForEntity(createTaskURL, request, Task.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public Task applyTask(Task task, boolean replace) {
-    try {
-      String url = applyTaskURL;
-      Map<String, String> requestParams = new HashMap<>();
-      requestParams.put("replace", Boolean.toString(replace));
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      final HttpHeaders headers = new HttpHeaders();
-      headers.setContentType(MediaType.APPLICATION_JSON);
-      HttpEntity<Task> entity = new HttpEntity<Task>(task, headers);
-      ResponseEntity<Task> response =
-          restTemplate.exchange(encodedURL, HttpMethod.PUT, entity, Task.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public List<ChangeLogVersion> getTaskChangeLog(String ref) {
-    try {
-      String url = changelogTaskURL.replace("{ref}", ref);
-
-      LOGGER.info("URL: " + url);
-
-      ResponseEntity<ChangeLogVersion[]> response =
-          restTemplate.getForEntity(url, ChangeLogVersion[].class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return Arrays.asList(response.getBody());
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public ResponseEntity<Void> deleteTask(String ref) {
-    try {
-      String url = deleteTaskURL.replace("{ref}", ref);
-
-      LOGGER.info("URL: " + url);
-      ResponseEntity<Void> response =
-          restTemplate.exchange(url, HttpMethod.DELETE, null, Void.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      return response;
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  /*
-   * **************************************
-   * WorkflowTemplate endpoints
-   * **************************************
-   */
-
-  public WorkflowTemplate getWorkflowTemplate(
-      String name, Optional<Integer> version, boolean withTasks) {
-    try {
-      String url = getWorkflowTemplateURL.replace("{name}", name);
-      Map<String, String> requestParams = new HashMap<>();
-      if (version.isPresent()) {
-        requestParams.put("version", version.toString());
-      }
-      requestParams.put("withTasks", Boolean.toString(withTasks));
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      ResponseEntity<WorkflowTemplate> response =
-          restTemplate.getForEntity(encodedURL, WorkflowTemplate.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public WorkflowTemplateResponsePage queryWorkflowTemplates(
-      Optional<Integer> queryLimit,
-      Optional<Integer> queryPage,
-      Optional<Direction> querySort,
-      Optional<List<String>> queryLabels,
-      Optional<List<String>> queryNames) {
-    try {
-      UriComponentsBuilder urlBuilder = UriComponentsBuilder.fromHttpUrl(queryWorkflowTemplateURL);
-      if (queryPage.isPresent()) {
-        urlBuilder.queryParam("page", Integer.toString(queryPage.get()));
-      }
-      if (queryLimit.isPresent()) {
-        urlBuilder.queryParam("limit", Integer.toString(queryLimit.get()));
-      }
-      if (querySort.isPresent()) {
-        urlBuilder.queryParam("sort", querySort.get());
-      }
-      if (queryLabels.isPresent()) {
-        urlBuilder.queryParam("labels", queryLabels.get());
-      }
-      if (queryNames.isPresent() && !queryNames.get().isEmpty()) {
-        urlBuilder.queryParam("names", queryNames.get());
-      }
-      URI encodedURI = urlBuilder.build().encode().toUri();
-
-      LOGGER.info("Query URL: " + encodedURI);
-
-      ResponseEntity<WorkflowTemplateResponsePage> response =
-          restTemplate.getForEntity(encodedURI, WorkflowTemplateResponsePage.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().getContent().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public WorkflowTemplate createWorkflowTemplate(WorkflowTemplate workflow) {
-    try {
-      LOGGER.info("URL: " + createWorkflowTemplateURL);
-
-      ResponseEntity<WorkflowTemplate> response =
-          restTemplate.postForEntity(createWorkflowTemplateURL, workflow, WorkflowTemplate.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public WorkflowTemplate applyWorkflowTemplate(WorkflowTemplate workflow, boolean replace) {
-    try {
-      String url = applyWorkflowTemplateURL;
-      Map<String, String> requestParams = new HashMap<>();
-      requestParams.put("replace", Boolean.toString(replace));
-
-      String encodedURL =
-          requestParams.keySet().stream()
-              .map(key -> key + "=" + requestParams.get(key))
-              .collect(Collectors.joining("&", url + "?", ""));
-
-      LOGGER.info("URL: " + encodedURL);
-
-      final HttpHeaders headers = new HttpHeaders();
-      headers.setContentType(MediaType.APPLICATION_JSON);
-      HttpEntity<WorkflowTemplate> entity = new HttpEntity<WorkflowTemplate>(workflow, headers);
-      ResponseEntity<WorkflowTemplate> response =
-          restTemplate.exchange(encodedURL, HttpMethod.PUT, entity, WorkflowTemplate.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      LOGGER.info("Content Response: " + response.getBody().toString());
-
-      return response.getBody();
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  public ResponseEntity<Void> deleteWorkflowTemplate(String name) {
-    try {
-      String url = deleteWorkflowTemplateURL.replace("{name}", name);
-
-      LOGGER.info("URL: " + url);
-      ResponseEntity<Void> response =
-          restTemplate.exchange(url, HttpMethod.DELETE, null, Void.class);
-
-      LOGGER.info("Status Response: " + response.getStatusCode());
-      return response;
-    } catch (RestClientException ex) {
-      LOGGER.error(ex.toString());
-      throw new BoomerangException(
-          ex,
-          HttpStatus.INTERNAL_SERVER_ERROR.value(),
-          ex.getClass().getSimpleName(),
-          "Exception in communicating with internal services.",
-          HttpStatus.INTERNAL_SERVER_ERROR);
-    }
   }
 }
